requirements.txt
Flask
mysql-connector-python
werkzeug
requests
reportlab
